pyproject.toml
[tool.poetry]
name = "slackbot_sample_size"
version = "0.1.0"
description = ""
authors = ["konumaru <konumaru1022@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
python-dotenv = "^0.20.0"
slack-bolt = "^1.13.0"
statsmodels = "^0.13.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
mypy = "^0.942"
pytest = "^7.1.2"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py39']
line-length = 80

[tool.isort]
line_length = 80
multi_line_output = 3
include_trailing_comma = true
known_local_folder=['config',]

[tool.flake8]
max-line-length = 80
max-complexity = 18
ignore = "E203,E266,W503,"

[tool.mypy]
python_version = "3.9"
no_strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests",]
filterwarnings = ["ignore::DeprecationWarning",]
